pyproject.toml
[project]
name = "maker-assistant"
version = "0.1.0"
description = "A multi-agent system to help me quickly find, prioritize, and complete creative projects."
authors = [
    {name = "Dallas Blowers"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.45.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "openai-agents (>=0.0.14,<0.0.15)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
